chore: move pkg/controller to controllers
