+++ cmd/manager/main.go
@@ -10,15 +10,15 @@ import (
 	"strings"
 
 	toolchainv1alpha1 "github.com/codeready-toolchain/api/api/v1alpha1"
+	"github.com/codeready-toolchain/member-operator/controllers/idler"
+	"github.com/codeready-toolchain/member-operator/controllers/memberstatus"
+	"github.com/codeready-toolchain/member-operator/controllers/nstemplateset"
+	"github.com/codeready-toolchain/member-operator/controllers/useraccount"
+	"github.com/codeready-toolchain/member-operator/controllers/useraccountstatus"
 	"github.com/codeready-toolchain/member-operator/pkg/apis"
 	"github.com/codeready-toolchain/member-operator/pkg/autoscaler"
 	"github.com/codeready-toolchain/member-operator/pkg/che"
 	"github.com/codeready-toolchain/member-operator/pkg/configuration"
-	"github.com/codeready-toolchain/member-operator/pkg/controller/idler"
-	"github.com/codeready-toolchain/member-operator/pkg/controller/memberstatus"
-	"github.com/codeready-toolchain/member-operator/pkg/controller/nstemplateset"
-	"github.com/codeready-toolchain/member-operator/pkg/controller/useraccount"
-	"github.com/codeready-toolchain/member-operator/pkg/controller/useraccountstatus"
 	"github.com/codeready-toolchain/member-operator/pkg/webhook/deploy"
 	"github.com/codeready-toolchain/member-operator/version"
 	"github.com/codeready-toolchain/toolchain-common/pkg/cluster"
